feat: start splitting apart expr_parser
