+++ crates/aiken-lang/src/parser/expr/list.rs
@@ -1,0 +1,29 @@
+use chumsky::prelude::*;
+
+use crate::{
+    expr::UntypedExpr,
+    parser::{error::ParseError, token::Token},
+};
+
+pub fn parser(
+    r: Recursive<'_, Token, UntypedExpr, ParseError>,
+) -> impl Parser<Token, UntypedExpr, Error = ParseError> + '_ {
+    just(Token::LeftSquare)
+        .ignore_then(r.clone().separated_by(just(Token::Comma)))
+        .then(choice((
+            just(Token::Comma).ignore_then(
+                just(Token::DotDot)
+                    .ignore_then(r.clone())
+                    .map(Box::new)
+                    .or_not(),
+            ),
+            just(Token::Comma).ignored().or_not().map(|_| None),
+        )))
+        .then_ignore(just(Token::RightSquare))
+        // TODO: check if tail.is_some and elements.is_empty then return ListSpreadWithoutElements error
+        .map_with_span(|(elements, tail), span| UntypedExpr::List {
+            location: span,
+            elements,
+            tail,
+        })
+}
